Fixing a bug in regen computation (#618)
* fixing a bug in regen computation

this PR fixes an issue in computation formula and avoids making overflow in hTorqueRegenMax value.

* fixing the indentations

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/src/middleware_motor/speed_torq_ctrl.c
@@ -356,14 +356,14 @@ int16_t SpdTorqCtrl_CalcTorqueReference(SpdTorqCtrlHandle_t * pHandle, MotorPara
 
          if (pHandle->motorType == DIRECT_DRIVE)
           
-        {
-          if (pHandle->pSPD->hIdcRegen)
-              {
-                if (pHandle->hCurrentTorqueRef == 0 && pHandle->pSPD->bActiveRegen)
-                {
+         {
+             if (pHandle->pSPD->hIdcRegen)
+             {
+                 if (pHandle->hCurrentTorqueRef == 0 && pHandle->pSPD->bActiveRegen)
+                 {
                       if (abs(pHandle->pSPD->hAvrMecSpeedUnit) > MIN_REGEN_SPEED)
-                        { 
-                          pHandle->pSPD->hTorqueRegenMax = (int16_t)(pHandle->pSPD->hIdcRegen * pHandle->hBusVoltage * 100)/(int16_t)(abs(pHandle->pSPD->hAvrMecSpeedUnit)*PI_/30);
+                      { 
+                          pHandle->pSPD->hTorqueRegenMax = (int16_t)((pHandle->pSPD->hIdcRegen * pHandle->hBusVoltage * 100)/(abs(pHandle->pSPD->hAvrMecSpeedUnit)*PI_/30));
                           if (pHandle->pSPD->hTorqueRegen < pHandle->pSPD->hTorqueRegenMax)
                           {
                             pHandle->pSPD->hTorqueRegen = pHandle->pSPD->hTorqueRegenMax;  
@@ -378,16 +378,15 @@ int16_t SpdTorqCtrl_CalcTorqueReference(SpdTorqCtrlHandle_t * pHandle, MotorPara
                           }
                   
                   
-                        }
+                      }
                       else
-                        {
+                      {
                           pHandle->pSPD->hTorqueRegen = 0;
-                        }
+                      }
                  }
                 
-                
               }
-            }  
+        }  
 
     }
     else
